% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sin01.R
\name{read_sin01}
\alias{read_sin01}
\title{Read Monitoring Site SIN01 (Tree Individual Survey) file}
\usage{
read_sin01(path, tidy = TRUE)
}
\arguments{
\item{path}{path to csv (TreeDataCSV2)}

\item{tidy}{logical. If \emph{true} (default), the year variable is made independent,
and the items that are repeatedly observed are stored in the variable.}
}
\description{
Read Monitoring Site SIN01 (Tree Individual Survey) file
}
\details{
Version 201906. Unknown and irregular value are replace to \emph{NA}.
}
\examples{
file <-
system.file("extdata/dummy-TreeGbh-2004-2018-ver2.transf.csv",
            package = "monit1000")
read_sin01(file, tidy = TRUE)
}
\seealso{
\url{http://www.biodic.go.jp/moni1000/findings/data/index_file.html}
}
